create get mapping for materials

+++ education-dao-hibernate/src/main/java/by/educ/ivan/education/dao/EducationalMaterialHibernateDao.java
@@ -50,7 +50,7 @@ public class EducationalMaterialHibernateDao implements EducationalMaterialDAO {
     public EducationalMaterial findEducationalMaterial(String educationalMaterialId) {
         EntityManager entityManager = entityManagerFactory.createEntityManager();
         try {
-            return entityManager.find(EducationalMaterial.class, Integer.valueOf(educationalMaterialId));
+            return entityManager.find(EducationalMaterial.class, Long.valueOf(educationalMaterialId));
         } finally {
             entityManager.close();
         }
